fix: clippy

+++ crates/glyim-codegen-llvm/src/codegen/types.rs
@@ -161,7 +161,7 @@ pub(crate) fn get_or_create_enum_struct_type<'ctx>(
     } else {
         // user-defined: use the stored payload array size
         if let Some((_, arr_ty)) = cg.enum_types.borrow().get(&enum_name) {
-            arr_ty.len() as u32
+            arr_ty.len()
         } else {
             8
         }
